+++ lib/glimmer/command_handler.rb
@@ -1,8 +1,13 @@
 module Glimmer
+  # rename to keyword handler
+
   module CommandHandler
     def can_handle?(parent, command_symbol, *args, &block)
       raise "must be implemented by a class"
     end
+
+    # TODO rename do_handle to just handle
+
     def do_handle(parent, command_symbol, *args, &block)
       raise "must be implemented by a class"
     end