+++ lib/glimmer/swt/command_handlers/widget_listener_command_handler.rb
@@ -5,6 +5,8 @@ require File.dirname(__FILE__) + "/../custom_widget"
 module Glimmer
   module SWT
     module CommandHandlers
+      # TODO rename to observation command handler (or keyword handler)
+
       class WidgetListenerCommandHandler
         include CommandHandler
 
@@ -14,18 +16,19 @@ module Glimmer
           widget_parentage = (parent.is_a?(GWidget) || parent.is_a?(CustomWidget))
           Glimmer.logger.debug "parent is a widget: #{widget_parentage}"
           return unless widget_parentage
-          Glimmer.logger.debug "on listener?: " + (command_symbol.to_s[0,3] == "on_").to_s
-          return unless command_symbol.to_s[0,3] == "on_"
-          Glimmer.logger.debug "command symbol is longer than 3: " + (command_symbol.to_s.length > 3).to_s
-          return unless command_symbol.to_s.length > 3
-          Glimmer.logger.debug "args are 0?: " + (args.size == 0).to_s
-          return unless args.size == 0
-          Glimmer.logger.debug "can add listener? " + (parent.can_add_listener?(command_symbol.to_s[3, command_symbol.to_s.length])).to_s
-          parent.can_add_listener?(command_symbol.to_s[3, command_symbol.to_s.length])
+          Glimmer.logger.debug "keyword starts with on_: #{command_symbol.to_s.start_with?('on_')}"
+          return unless command_symbol.to_s.start_with?('on_')
+          Glimmer.logger.debug "args are empty?: #{args.empty?}"
+          return unless args.empty?
+          Glimmer.logger.debug "block exists?: #{!block.nil?}"
+          return unless !block.nil?
+          result = parent.can_handle_observation_request?(command_symbol.to_s)
+          Glimmer.logger.debug "can add listener? #{result}"
+          result
         end
 
         def do_handle(parent, command_symbol, *args, &block)
-          parent.add_listener(command_symbol.to_s[3, command_symbol.to_s.length], &block)
+          parent.handle_observation_request(command_symbol.to_s, &block)
           ListenerParent.new #TODO refactor and move to models
         end